+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-13T07:48:44.054Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-13T09:21:05.843Z")
     r"""The date and time the resource was last updated"""