+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9Y30QAKVWF6C0H2GEGAY39J"
+        "01K9Y89TFT967PW4EQNCK2V1S4"
     )
     r"""The unique identifier of the data source"""