+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -215,5 +215,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-13T07:48:44.390Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-13T09:21:06.136Z")
     r"""The date and time the resource was last updated"""