+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9Y30QB0YF5VKBHXG6ZB0A6V"
+        "01K9Y89TFZWV55DD8B5VCQ0GJR"
     )
     r"""The unique identifier of the data source"""