+++ packages/orq-rc/RELEASES.md
@@ -3979,3 +3979,13 @@ Based on:
 - [python v4.0.0-rc.28] packages/orq-rc
 ### Releases
 - [PyPI v4.0.0-rc.28] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.28 - packages/orq-rc
+
+## 2025-11-13 09:22:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.656.1 (2.753.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.0-rc.29] packages/orq-rc
+### Releases
+- [PyPI v4.0.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.29 - packages/orq-rc